Infrastructure/python-etl/requirements.txt
yfinance==0.2.66
psycopg2-binary==2.9.11